adds logging messages for experiment service

+++ src/machine_learning/training/batch_training_service.py
@@ -76,8 +76,8 @@ class BatchTrainingService(TrainingService[TInput, TTarget, TModel], ABC):
 
         batch_size: int = len(dataset) if self.__batch_size is None else self.__batch_size
         training_context: TrainingContext[TModel] = TrainingContext[TModel](model=model, scores={objective: [] for objective in objective_functions.keys()}, _primary_objective=primary_objective, current_epoch=0, current_iteration=0)
-        training_size: int = len(dataset) * self.__training_dataset_size_ratio
-        validation_size: int = len(dataset) - training_size
+        training_size: int = int(len(dataset) * self.__training_dataset_size_ratio)
+        validation_size: int = int(len(dataset) - training_size)
 
         if validation_dataset is None: 
             training_dataset, validation_dataset = random_split(dataset, [training_size, validation_size])
